+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning 2.0.0](https://semver.org/spec/
 ### Fixed
 
 - Comment count while getting posts
+- JWT refresh tokens
+- Account Recovery (forgot password feature)
+- Skills Case Normalization
 
 ## [2.0.0] - 2026-06-29